+++ src/main/kotlin/com/torben/webshop/model/PersonRepository.kt
@@ -1,6 +1,5 @@
-package com.torben.webshop
+package com.torben.webshop.model
 
-import org.springframework.boot.autoconfigure.security.SecurityProperties
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import java.util.Optional
