+++ src/test/kotlin/com/torben/webshop/PersonRepositoryTest.kt
@@ -1,5 +1,7 @@
 package com.torben.webshop
 
+import com.torben.webshop.model.PersonEntity
+import com.torben.webshop.model.PersonRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
